[process-agent] Move deprecated flag handling to separate file
